Refactor entire mapper to build text correctly
As the text element must have the description last, with other details
prepended to it, and since the text is generated from different parts of
the mapper which could not hold state, an internal mapper has been added
to encapsulate the process of mapping a single request. Multiple threads
accessing the (outer) mapper would trigger individual inner mappers, and
so the inner mappers may hold state thread-safely. In this way, the text
can be held as an instance variable and manipulated through the mapping
process, and then added to the template parameters prior to 'mustaching'.