Encode the referenceRange text for XML
Without encoding, range text such as "< 10" would be copied literally into the output XML,
making it badly-formed. Instead, this example should be "&lt; 10"

I saw two immediate options:
1. To use apache commons escaping utils
2. To convert the string format constants into mustache templates, where escaping is given
 out of the box.

I opted for the latter for consistency with the rest of the codebase.

This included introducing a new way to create "ad hoc" Mustache objects that can be used
just like the other templates we're creating from files.