Set OriginalBase64 to No for a Skeleton message
This value according to the GP2GP spec is:
> The CCLM process sends attachments as base64 encoded streams.
> If the “original” for the attachment was already base64
> this is not done, but the “OriginalBase64” value is Yes.

In the Skeleton message case, the original was XML, so No is an
appropriate value.

In the examples provided within the spec, the Skeleton messages
are also with No.

> Filename="EEB93039-4285-4937-AEDB-18844C14DC9inps.co.ukVision3.gzip"
> ContentType=text/xml Compressed=No LargeAttachment=No OriginalBase64=No
> Length=302448 DomainData="X-GP2GP-Skeleton: Yes"

SystmOne also provide a value of No when sending their Skeleton message out.