Add PDS data within Wiremock migratestructuredrecord stub response
We believe this will reduce the time spend hand crafting and deploying
specific wiremock stubs when testing in our INT Path to Live environment.

Follows on from 4b2254b37e697312a22ee06d8f3fd649de021545 which was believed
to have a similar effect, but it would appear that EMIS/TPP validate the
patient details which come back. The only way I can see to avoid that failure
in a pain-free way is to integrate with PDS directly, and inject the correct
patients details into the response so that EMIS won't error.

This only adds PDS data to `correctPatientNoDocsStructuredRecordResponse.json`
but the same pattern could be expanded across all responses in a follow up PR.