Introduce Mutation Testing quality check GitHub Action (#790)
* Introduce Mutation Testing quality check GitHub Action

This workflow will identify code within a PR which isn't fully
covered with tests using a process called Mutation testing.

This process isn't perfect, but hopefully the feedback it does
identify can start a conversation about where tests are missing.

* Remove config folder which crashes pitest

The pitest git integration tries to automatically find the git repo
root directory. In doing so, the presence of a directory called config
causes it to crash.