NIAD-3150: Add confidentialityCode to MedicationStatements.
Add functionality to provide `confidentialityCode` for both `PLAN` and `ORDER` when mapping `MedicationRequests` which contain a `NOPAT` `meta.security` tag.
Add Unit Tests to MedicationStatementMapperTest to test mapping of confidentialityCode when present or not present for both `supplyPrescribe` and `supplyAuthorise`
Update MedicationStatementMapper to inject ConfidentialityService and to add parameter for `confidentialityCode` to MedicationStatementTemplateParameters.
Update mustache template to include mapping for confidentialityCode for both `PLAN` and `ORDER`.
Update injection of ConfidentialityService in related test files.
Update CHANGELOG.md