NIAD-1615: `ProcedureRequest` displays 'Earliest Recall Date' text when `occurrencePeriod` has only a start date.
Created explict tests to cover the scenarios of when the occurrence period in the input json `ProcedureRequest` resource, contains both and start and an end date, and when it only contains a start date.
Removed test files and related entries in parameterized tests as now covered by the explicit tests.
Updated functionality to only produce the "Earliest Recall Date" text in the XML `ProcedureRequest` when the input json `ProcedureRequest.occurrencePeriod` has both a start and an end date