NIAD-1616: when `ProcedureRequest` requester is `Device`, `null` is added to text when `manufacturer` is not present.
Add `input-bundle-no-manufacturer.json` to represent a device without a manufacturer for unit testing.
Add unit test to ensure that when the `device` has `device type text` and `manufacturer`, both are shown in the `"Recall Device:"` text.
Add unit test to ensure that when the device has `device type text` and no `manufacturer`, only `device type text` is shown in the `"Recall Device:"` text.
Add functionality to `DiaryPlanStatementMapper` when formatting the `"Recall Device:"` text that the missing the `manufacturer` is handled and that `null` values are not displayed in the text.