Allow sending an ehrComposition without an author/Participant2
Having done some testing with other GP2GP implementations
we've identified that other systems will accept this and
generate a fake placeholer Dr within their system when ingesting
the record.

While the GP Connect spec insists that an Encounter has a participant
of type REC, it is preferable to send a GP2GP extract than fail the
tranfer completely.