Improve realism of Observation (Filing comment) tests
These test resources aren't valid GP Connect filing comments because
the "specimen" field is not a valid field for a filing comment.

The side effect of having the specimen field populated twice is that
the expected XML was generating two NarrativeStatements with the text
"This is some random free text".