NIAD-3222: GP2GP adaptor generates an empty `CompoundStatement/code` which fails the transfer.
Replace String.Empty when preparing the code element mapping ObservationMapper so that an EhrMapperException is thrown containing the resource identifier.
Add test to ensure this exception is found and contains the resource id.
Update NOPAT test and associated resource file which previously did not contain a code element
Update CHANGELOG.md